% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_print_glm.R
\name{apa_print_glm}
\alias{apa_print_glm}
\title{Add odds ratios and model fit statistics to a glm object}
\usage{
apa_print_glm(model)
}
\arguments{
\item{model}{a glm() object generated from a binomial logistic regression}
}
\value{
A papaja apa_print object
}
\description{
Default functions for logistic regressions do not output all commonly required
statistics for reporting. This function adds odds ratios and associated
confidence intervals and relevant model fit statistics.
}
\details{
Qualtrics exports data with bloat, this function is an opinionated wrapper of
readr::read_csv() which does some initial cleaning.
}
