% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_names.R
\name{cat_names}
\alias{cat_names}
\title{Returns unquoted column names for scripting}
\usage{
cat_names(df, pattern = NULL, index = NULL, sep = "\\n")
}
\arguments{
\item{df}{a dataframe or tibble}

\item{pattern}{an optional regex for selection of column names}

\item{index}{an optional vector of column indices for selection}

\item{sep}{seperator for column names, suggestions include ", ", " + ", and combinations with newline.}
}
\value{
An unquoted string of column names
}
\description{
Returns unquoted column names for scripting
}
\examples{
df <-  data.frame(A = 0L, B= 0L, C = 0L, test1 = 0L, test2 = 0L)
cat_names(df)
cat_names(df, sep = ',')
cat_names(df, index =2:3, sep = '+')
cat_names(df, pattern = "test", sep = '+')
}
