% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country codes}
\format{A dataframe with ten variables: country, alpha_2,
  alpha_3,  contry_code,  iso_3166_2,  region,
  sub_region,  region_code,  sub_region_code,
  qualtrics.}
\usage{
countries
}
\description{
Country names, regions and ISO 3166 codes, paired with the reference number in the default country question in Qualtrics
}
\keyword{datasets}
