% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_rename.R
\name{meta_rename}
\alias{meta_rename}
\title{Replaces names by reference to a metadata dataframe}
\usage{
meta_rename(df, metadata, old, new)
}
\arguments{
\item{df}{A data.frame object for renaming.}

\item{metadata}{A data.frame object containing the key-value pair}

\item{old}{An unquoted string of the key column name in the metadata object}

\item{new}{An unquoted string of the value column name in the metadata object}
}
\value{
`data.frame` object with new names as per key-value pairs.
}
\description{
`meta_rename` allows fast replacement of variable names by key-value pairing with a metadata file.
}
\examples{

testdata <-  tibble(a = 0L,
b = 0L,
c = 0L)

meta <-  tibble(DatasetPositon = 1:3,
                VariableID =	letters[1:3],
                VariableType = "integer",
                VariableLabel = c("item1", "item2", "item3"),
                ItemText = c("This is item 1",
                             "This is item 2",
                             "This is item 3")
)

meta_rename(testdata, metadata = meta, old = VariableID, new = VariableLabel)

}
