% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtr_countries.R
\name{qtr_countries}
\alias{qtr_countries}
\title{Recodes qualtrics default contry codes}
\usage{
qtr_countries(countrycode)
}
\arguments{
\item{countrycode}{A number referring to a default reference in the qualtrics country variable.}
}
\value{
The Name of the corresponding country
}
\description{
Recodes qualtrics default contry codes
}
