% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{Read qualtrics data csv}
\usage{
read_qualtrics(file, legacy = FALSE)
}
\arguments{
\item{file}{a csv of data exported from qualtrics}

\item{legacy}{a Logical indicating if the export used the legacy exporter}
}
\value{
A tibble of data
}
\description{
Qualtrics exports data with bloat, this function is an opinionated wrapper of
readr::read_csv() which does some initial cleaning.
}
