% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_fees.R
\name{recruit_fees}
\alias{recruit_fees}
\title{Calculate recruitment costs in MTurk, ProlificAc and Turkprime}
\usage{
recruit_fees(n, duration, pay)
}
\arguments{
\item{n}{Target sample}

\item{duration}{time (in minutes) to complete study}

\item{pay}{rate of pay per participant in USD}
}
\value{
A dataframe of calculated costs
}
\description{
Calculate recruitment costs in MTurk, ProlificAc and Turkprime
}
