% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_n.R
\name{required_n}
\alias{required_n}
\title{Attrition and incompleteness calculator}
\usage{
required_n(final, attrition = 0.1, trials = 1)
}
\arguments{
\item{final}{integer target n}

\item{attrition}{The decimal rate of attrition from 0 (0%) to 1 (100%)}

\item{trials}{An integer number of measurement points over which attrition occurs
(i.e. after the first measurement). If calculating loss to missingness, specify as 1.}
}
\value{
An integer of the required starting sample
}
\description{
A function for calculating the required starting
 sample to correct for attrition and incomplete data.
}
